CONTRIB-9347: Fix send_notifications (#609)
