+++ classes/task/send_recording_ready_notification.php
@@ -25,7 +25,7 @@ use html_writer;
  * @copyright 2021 Andrew Lyons <andrew@nicols.co.uk>
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class send_recording_ready_notification extends send_notification {
+class send_recording_ready_notification extends base_send_notification {
     /**
      * Get the notification type.
      *
